% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_data.R
\name{create_struc_data}
\alias{create_struc_data}
\title{Creates random STRUCTURE data of the format required for plotting}
\usage{
create_struc_data(num_indiv = 500, num_clusters = 10)
}
\arguments{
\item{num_indiv}{The number of individuals to produce, defaults to 500.}

\item{num_clusters}{The number of clusters to produce, defaults to 10.}
}
\value{
A tibble with three columns: Individual (character),
Cluster (integer) and q (double).
}
\description{
The function creates a tibble with three columns: Individual, Cluster, q.
Users can control the number of individuals and clusters created and then
test using the plotting functions.
}
\examples{
create_struc_data()
struc_plot(create_struc_data())

}
