% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_data.R
\name{struc_reorder}
\alias{struc_reorder}
\title{Reorder structure data before plotting}
\usage{
struc_reorder(struc_data)
}
\arguments{
\item{struc_data}{The data frame or tibble to reorder. The function expects
that the column names of the data will be Individual, Cluster and q. The
Individual column will automatically coerced to character.}
}
\value{
The data frame or tibble reordered.
}
\description{
The function takes in a data frame or tibble and returns it reordered by the
clusters per individual with descending values of q.
}
